Refactor ff-helper.ts and add getFFLocalServices function
This commit moves getFFPath function towards the end of the file and introduces a new function named getFFLocalServices. Functionality of getFFPath remains the same. The newly added getFFLocalServices function is aimed at getting the local services in the FactoryFix directory. It filters the local services based on whether they are built using nest.js or not. This will be helpful in selecting a local service for injection of a debug configuration.