Refactor folder selection from helper to consumer script
To offer more flexibility on folder selection, we have moved the `arg` selection from the `getFFLocalServices` method in the helper file to the consumer script. Now, instead of selecting a folder in the helper method, it returns all valid folders and the selection is done in the `ff-generate-jwt-token` script instead.