Refactor ff-merged-to-prod script to use FFService
The ff-merged-to-prod script is updated to use FFService and bin-helper modules. Previously, the script directly accessed system directories which could become a bottleneck with more complex project structures. Now, using FFService simplifies the fetching of project paths and increases the modularity of the code. Implementation of bin-helper assures the code not to use hardcoded paths and enhances scalability. Additionally, the notification calls have also been refactored to be synchronous.