Rename and refactor script for extracting test environment variables
The script previously known as 'ff-convert-yaml-env-for-test.ts' has been renamed to 'ff-extract-test-envs.ts'. The functionality of the script has been modified to enhance ease-of-use and functionality. Instead of just converting environment variables into a different format, the script now also fetches local services, lets the user select a service, extracts the environment variables from the chosen service, and stores them in a more user-friendly format. This change aims to improve the utility of the script for users who want to quickly and efficiently extract environment variables for testing purposes.